fix: remove duplicated code of `useEffect`
