+++ web/src/pages/desktop/screen/h264-direct.tsx
@@ -1,9 +1,8 @@
 import { useEffect, useRef } from 'react';
 import clsx from 'clsx';
-import { useAtom, useAtomValue } from 'jotai';
+import { useAtomValue } from 'jotai';
 
 import * as api from '@/api/stream.ts';
-import * as storage from '@/lib/localstorage.ts';
 import { mouseStyleAtom } from '@/jotai/mouse';
 import { videoScaleAtom } from '@/jotai/screen.ts';
 
@@ -11,17 +10,10 @@ import DirectWorker from './direct.worker.ts?worker';
 
 export const H264Direct = () => {
   const mouseStyle = useAtomValue(mouseStyleAtom);
-  const [videoScale, setVideoScale] = useAtom(videoScaleAtom);
+  const videoScale = useAtomValue(videoScaleAtom);
 
   const canvasRef = useRef<HTMLCanvasElement>(null);
   const workerRef = useRef<Worker | null>(null);
-
-  useEffect(() => {
-    const scale = storage.getVideoScale()
-    if (scale) {
-      setVideoScale(scale)
-    }
-  }, [setVideoScale])
 
   useEffect(() => {
     if (!canvasRef.current) {
@@ -66,14 +58,10 @@ export const H264Direct = () => {
       <canvas
         id="screen"
         ref={canvasRef}
-        className={clsx('block min-h-[480px] min-w-[640px] select-none', mouseStyle)}
+        className={clsx('block min-h-[480px] min-w-[640px] select-none origin-center max-w-full max-h-[calc(100%_-_75px)] object-scale-down', mouseStyle)}
         style={{
           transform: `scale(${videoScale})`,
-          transformOrigin: 'center',
-          maxWidth: '100%',
-          maxHeight: 'calc(100% - 75px)',
-          objectFit: 'scale-down'
-      }}
+        }}
       ></canvas>
     </div>
   );
