+++ web/src/pages/desktop/screen/h264-webrtc.tsx
@@ -7,11 +7,10 @@ import * as api from '@/api/stream.ts';
 import { VideoStatus } from '@/types';
 import { mouseStyleAtom } from '@/jotai/mouse.ts';
 import { videoStatusAtom, videoVolumeAtom, videoScaleAtom } from '@/jotai/screen.ts';
-import * as storage from '@/lib/localstorage.ts'
 
 export const H264Webrtc = () => {
   const mouseStyle = useAtomValue(mouseStyleAtom);
-  const [videoScale, setVideoScale] = useAtom(videoScaleAtom);
+  const videoScale = useAtomValue(videoScaleAtom);
   const setVideoStatus = useSetAtom(videoStatusAtom);
   const [volume, setVolume] = useAtom(videoVolumeAtom);
 
@@ -302,13 +301,6 @@ export const H264Webrtc = () => {
     audioRef.current.volume = volume / 100;
   }, [volume]);
 
-  useEffect(() => {
-    const scale = storage.getVideoScale()
-    if (scale) {
-      setVideoScale(scale)
-    }
-  }, [setVideoScale])
-
   return (
     <Spin size="large" tip="Loading" spinning={isLoading}>
       <div className="flex h-screen w-screen items-start justify-center xl:items-center">
@@ -316,12 +308,11 @@ export const H264Webrtc = () => {
           id="screen"
           ref={videoRef}
           className={clsx(
-            'block max-h-full min-h-[480px] min-w-[640px] max-w-full select-none object-scale-down',
+            'block max-h-full min-h-[480px] min-w-[640px] max-w-full select-none object-scale-down origin-center',
             mouseStyle
           )}
           style={{
             transform: `scale(${videoScale})`,
-            transformOrigin: 'center'
           }}
           muted
           autoPlay
