+++ web/src/pages/desktop/screen/mjpeg.tsx
@@ -1,30 +1,20 @@
-import { useEffect } from 'react';
 import { Image } from 'antd';
 import clsx from 'clsx';
-import { useAtom, useAtomValue } from 'jotai';
+import { useAtomValue } from 'jotai';
 
 import MonitorXIcon from '@/assets/images/monitor-x.svg';
 import { getBaseUrl } from '@/lib/service.ts';
 import { mouseStyleAtom } from '@/jotai/mouse.ts';
 import { videoScaleAtom } from '@/jotai/screen.ts';
-import * as storage from '@/lib/localstorage.ts'
 
 export const Mjpeg = () => {
   const mouseStyle = useAtomValue(mouseStyleAtom);
-  const [videoScale, setVideoScale] = useAtom(videoScaleAtom);
-
-  useEffect(() => {
-    const scale = storage.getVideoScale()
-    if (scale) {
-      setVideoScale(scale)
-    }
-  }, [setVideoScale])
+  const videoScale = useAtomValue(videoScaleAtom);
 
   return (
-    <div className="flex h-screen w-screen items-start justify-center xl:items-center"
+    <div className="flex h-screen w-screen items-start justify-center xl:items-center origin-center"
          style={{
             transform: `scale(${videoScale})`,
-            transformOrigin: 'center'
           }}>
       <Image
         id="screen"
